perf(jsx): move y axis function to utils to be reuable across charts
